Modified README with video link, demo link, and Lighthouse screenshots:
Added link to a video showcasing the game and included a link
to the live demo hosted on GitHub pages.
Also included screenshots from Lighthouse, an open-source,
automated tool for improving the performance, quality,
and correctness of web apps, which show the performance
and accessibility scores.

No changes were made to the code,
this was solely a modification to the README file.

Commit made by Andrés Valencia.